backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Marshmallow==0.15.0
marshmallow==3.21.3
marshmallow-sqlalchemy==0.29.0
Werkzeug==2.3.7
python-dotenv==1.0.0
psycopg[binary]==3.2.10
requests==2.31.0